refactor: change name of dictionary loader
