+++ packages/next/src/config-dir/DictionaryManager.ts
@@ -1,6 +1,6 @@
 import { standardizeLocale } from 'generaltranslation';
 import { DictionaryObject } from 'gt-react/internal';
-import resolveDictionaryLoader from '../loaders/resolveDictionary';
+import resolveDictionaryLoader from '../loaders/resolveDictionaryLoader';
 import { customLoadDictionaryWarning } from '../errors/createErrors';
 
 /**
